Extract grid components
