--- src/Spreadsheet/FixLines/FixLinesView.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-const FixLinesView = ({ type, style }) => {
-  const className = type === 'rows' ? 'fixed-rows' : 'fixed-columns'
-  return <div className={className} style={style} />;
-};
-
-export default FixLinesView;
