--- src/Spreadsheet/FixLines/index.js
@@ -1,1 +1,0 @@
-export { default } from './FixLines';
