--- src/Spreadsheet/SpecialCellEmptyArea/SpecialCellEmptyArea.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import SpreadsheetCell from '../SpreadsheetCell';
-
-const SpecialCellEmptyArea = props => <SpreadsheetCell className="special-cell-empty-area" {...props} />;
-
-export default SpecialCellEmptyArea;
