--- src/Spreadsheet/SpecialCellEmptyArea/index.js
@@ -1,1 +1,0 @@
-export { default } from './SpecialCellEmptyArea';
