+++ src/grid/MergedCell/utils.test.js
