+++ src/grid/utils.js
@@ -1,0 +1,12 @@
+export const visibleRangesFilter = ({
+  fixRows,
+  fixColumns,
+  visibleRows,
+  visibleColumns
+}) => mergedRange => {
+  return mergedRange.start.row < visibleRows[fixRows] || mergedRange.start.column < visibleColumns[fixColumns] ||
+      (mergedRange.start.row <= visibleRows[visibleRows.length - 1] &&
+          mergedRange.start.column <= visibleColumns[visibleColumns.length - 1] &&
+          mergedRange.end.row >= visibleRows[fixRows] &&
+          mergedRange.end.column >= visibleColumns[fixColumns]);
+};
